+++ models/personal_goal.py
@@ -9,11 +9,11 @@ class Personal_goal(db.Model, SerializerMixin):
     id = db.Column(db.Integer, primary_key=True)
     
     name = db.Column(db.String, nullable=False)
-    saving_target = db.Column(db.Float, default=0, nullable=False)
+    saving_target = db.Column(db.Float(2), default=0, nullable=False)
     start_timeframe = db.Column(db.DateTime, server_default=db.func.now())
     end_timeframe = db.Column(db.DateTime, nullable=False)
     emoji = db.Column(db.String)
-    amount_saved = db.Column(db.Float, default=0)
+    amount_saved = db.Column(db.Float(2), default=0)
     
     user_id = db.Column(db.Integer, db.ForeignKey("users.id"), nullable=False)
     insights = db.relationship("Insight", cascade="all, delete", backref="personal_goal")
